requirements.txt
requests==2.14.2
pymongo==3.5.1
websocket_client==0.44.0
